+++ io-hotmoka-beans-api/src/main/java/io/hotmoka/beans/api/values/LongValue.java
@@ -19,15 +19,15 @@ package io.hotmoka.beans.api.values;
 import io.hotmoka.annotations.Immutable;
 
 /**
- * A {@code byte} value stored in blockchain.
+ * A {@code long} value stored in blockchain.
  */
 @Immutable
-public interface ByteValue extends StorageValue {
+public interface LongValue extends StorageValue {
 
 	/**
-	 * Yields the byte inside this object.
+	 * Yields the long inside this value.
 	 * 
-	 * @return the byte
+	 * @return the long
 	 */
-	byte getValue();
+	long getValue();
 }
