+++ io-hotmoka-beans-api/src/main/java/io/hotmoka/beans/api/values/ShortValue.java
@@ -19,15 +19,15 @@ package io.hotmoka.beans.api.values;
 import io.hotmoka.annotations.Immutable;
 
 /**
- * A {@code byte} value stored in blockchain.
+ * A {@code short} value stored in blockchain.
  */
 @Immutable
-public interface ByteValue extends StorageValue {
+public interface ShortValue extends StorageValue {
 
 	/**
-	 * Yields the byte inside this object.
+	 * Yields the short inside this value.
 	 * 
-	 * @return the byte
+	 * @return the short
 	 */
-	byte getValue();
+	short getValue();
 }
